// src/components/Text/types.ts
/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
export enum TextColor {
  /**
   * Primary black text color
   */
  PRIMARY = 'primary',

  /**
   * Secondary dark gray text color
   */
  SECONDARY = 'secondary',

  /**
   * Tertiary color for metadata
   */
  METADATA = 'metadata',

  /**
   * For use on darker backgrounds
   */
  WHITE = 'white',

  /**
   * For use in error messages.
   */
  ERROR = 'error',

  /**
   * For use in interaction elements which have become disabled.
   */
  DISABLED = 'disabled',
}

export enum TextSize {
  /**
   * Large headers. Font-size 2.4rem, line-height 2.8rem.
   */
  XXLARGE = 'xxLarge',

  /**
   * Group headers. Font-size 2.2rem, line-height 2.8rem.
   */
  XLARGE = 'xLarge',

  /**
   * Promo headlines. Font-size 1.8rem, line-height 2.4rem.
   */
  LARGE = 'large',

  /**
   * Body text, default size. Font-size 1.5rem, line-height 2.0rem.
   */
  MEDIUM = 'medium',

  /**
   * Secondary text. Font-size 1.4rem, line-height 2.0rem.
   */
  MEDIUM_SUB = 'mediumSub',

  /**
   * Metadata. Font-size 1.2rem, line-height 1.6rem.
   */
  SMALL = 'small',

  /**
   * Very limited use. Font-size 1.0rem, line-height 1.6rem.
   */
  XSMALL = 'xSmall',
}